refactor: refactor UI components into separate files
- Moved composable UI components such as InfoRow, TemperamentChips, EmptyFavoritesState, TypeCard, BackgroundDecorations, and all ProfileScreen subcomponents into their own files
- Updated imports and removed inlined component definitions from main screen files.
