+++ feature/details/src/main/java/com/example/details/TemperamentChips.kt
@@ -1,0 +1,43 @@
+package com.example.details
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.material3.AssistChip
+import androidx.compose.material3.AssistChipDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalLayoutApi::class)
+@Composable
+fun TemperamentChips(
+    temperament: String,
+    modifier: Modifier = Modifier
+) {
+    val temperaments = temperament.split(",").map { it.trim() }.filter { it.isNotEmpty() }
+
+    FlowRow(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        temperaments.forEach { trait ->
+            AssistChip(
+                onClick = {  },
+                label = {
+                    Text(
+                        text = trait,
+                        style = MaterialTheme.typography.labelMedium
+                    )
+                },
+                colors = AssistChipDefaults.assistChipColors(
+                    containerColor = MaterialTheme.colorScheme.secondaryContainer,
+                    labelColor = MaterialTheme.colorScheme.primary
+                )
+            )
+        }
+    }
+}
