+++ feature/onboarding/src/main/java/com/example/onboarding/OnboardingScreen.kt
@@ -229,159 +229,3 @@ fun OnboardingScreen(
     }
 }
 
-@Composable
-fun TypeCard(
-    painter: Painter,
-    label: String,
-    petType: PetType,
-    isSelected: Boolean,
-    onClick: () -> Unit
-) {
-    val scale by animateFloatAsState(
-        targetValue = if (isSelected) 1.1f else 1f,
-        animationSpec = spring(
-            dampingRatio = Spring.DampingRatioMediumBouncy,
-            stiffness = Spring.StiffnessLow
-        ),
-        label = "card_scale"
-    )
-
-    val elevation by animateDpAsState(
-        targetValue = if (isSelected) 16.dp else 8.dp,
-        animationSpec = spring(
-            dampingRatio = Spring.DampingRatioMediumBouncy,
-            stiffness = Spring.StiffnessLow
-        ),
-        label = "card_elevation"
-    )
-
-    Card(
-        modifier = Modifier
-            .widthIn(min = 140.dp, max = 160.dp)
-            .aspectRatio(1f)
-            .scale(scale)
-            .clickable { onClick() },
-        elevation = CardDefaults.cardElevation(defaultElevation = elevation),
-        shape = RoundedCornerShape(20.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = if (isSelected) {
-                MaterialTheme.colorScheme.primaryContainer
-            } else {
-                MaterialTheme.colorScheme.surface
-            }
-        ),
-        border = if (isSelected) {
-            androidx.compose.foundation.BorderStroke(
-                2.dp,
-                MaterialTheme.colorScheme.primary
-            )
-        } else null
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            val emojiScale by animateFloatAsState(
-                targetValue = if (isSelected) 1.2f else 1f,
-                animationSpec = spring(
-                    dampingRatio = Spring.DampingRatioMediumBouncy,
-                    stiffness = Spring.StiffnessLow
-                ),
-                label = "emoji_scale"
-            )
-
-            Icon(
-                painter = painter,
-                contentDescription = label,
-                modifier = Modifier.size(48.dp).scale(emojiScale),
-                tint = MaterialTheme.colorScheme.primary
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Text(
-                text = label,
-                style = MaterialTheme.typography.titleMedium.copy(
-                    fontWeight = FontWeight.SemiBold
-                ),
-                textAlign = TextAlign.Center,
-                color = if (isSelected) {
-                    MaterialTheme.colorScheme.onPrimaryContainer
-                } else {
-                    MaterialTheme.colorScheme.onSurface
-                }
-            )
-
-            Spacer(modifier = Modifier.height(4.dp))
-
-            // Selection indicator
-            if (isSelected) {
-                Spacer(modifier = Modifier.height(8.dp))
-
-                Card(
-                    modifier = Modifier.size(24.dp),
-                    shape = CircleShape,
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.primary
-                    )
-                ) {
-                    Box(
-                        modifier = Modifier.fillMaxSize(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(
-                            text = "✓",
-                            color = Color.White,
-                            fontSize = 12.sp,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun BackgroundDecorations() {
-    // Floating decoration circles
-    Box(modifier = Modifier.fillMaxSize()) {
-        // Top left circle
-        Card(
-            modifier = Modifier
-                .size(100.dp)
-                .offset((-30).dp, 50.dp),
-            shape = CircleShape,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.1f)
-            )
-        ) {}
-
-        // Top right circle
-        Card(
-            modifier = Modifier
-                .size(60.dp)
-                .offset(20.dp, 120.dp)
-                .align(Alignment.TopEnd),
-            shape = CircleShape,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.secondary.copy(alpha = 0.1f)
-            )
-        ) {}
-
-        // Bottom left circle
-        Card(
-            modifier = Modifier
-                .size(80.dp)
-                .offset((-20).dp, (-40).dp)
-                .align(Alignment.BottomStart),
-            shape = CircleShape,
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.tertiary.copy(alpha = 0.1f)
-            )
-        ) {}
-    }
-}
