+++ feature/profile/src/main/java/com/example/profile/components/EditTextDialog.kt
@@ -1,0 +1,59 @@
+package com.example.profile.components
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.example.profile.R.string
+
+@Composable
+fun EditTextDialog(
+    title: String,
+    currentValue: String,
+    placeholder: String,
+    onDismiss: () -> Unit,
+    onConfirm: (String) -> Unit,
+    singleLine: Boolean = true
+) {
+    var textValue by remember { mutableStateOf(currentValue) }
+
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(title) },
+        text = {
+            OutlinedTextField(
+                value = textValue,
+                onValueChange = { textValue = it },
+                placeholder = { Text(placeholder) },
+                singleLine = singleLine,
+                modifier = Modifier.fillMaxWidth(),
+                maxLines = if (singleLine) 1 else 3
+            )
+        },
+        confirmButton = {
+            Button(
+                onClick = {
+                    if (textValue.isNotBlank()) {
+                        onConfirm(textValue.trim())
+                    }
+                }
+            ) {
+                Text(stringResource(string.save))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(string.cancel))
+            }
+        }
+    )
+}
