+++ feature/profile/src/main/java/com/example/profile/components/ThemeModeDialog.kt
@@ -1,0 +1,72 @@
+package com.example.profile.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Brightness4
+import androidx.compose.material.icons.filled.DarkMode
+import androidx.compose.material.icons.filled.LightMode
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.example.preferences.ThemeMode
+import com.example.profile.R.string
+
+@Composable
+fun ThemeModeDialog(
+    currentThemeMode: ThemeMode,
+    onDismiss: () -> Unit,
+    onSelect: (ThemeMode) -> Unit
+) {
+    // TODO: Improve Dialog UI
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(string.choose_theme)) },
+        text = {
+            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
+                Text(stringResource(string.select_preferred_theme))
+                Spacer(modifier = Modifier.height(8.dp))
+
+                ThemeMode.entries.forEach { mode ->
+                    FilterChip(
+                        selected = currentThemeMode == mode,
+                        onClick = { onSelect(mode) },
+                        label = {
+                            Text(
+                                when (mode) {
+                                    ThemeMode.LIGHT -> stringResource(string.light_mode)
+                                    ThemeMode.DARK -> stringResource(string.dark_mode)
+                                    ThemeMode.SYSTEM -> stringResource(string.system_default)
+                                }
+                            )
+                        },
+                        leadingIcon = {
+                            val icon = when (mode) {
+                                ThemeMode.LIGHT -> Icons.Default.LightMode
+                                ThemeMode.DARK -> Icons.Default.DarkMode
+                                ThemeMode.SYSTEM -> Icons.Default.Brightness4
+                            }
+                            Icon(icon, null, Modifier.size(18.dp))
+                        },
+                        modifier = Modifier.fillMaxWidth()
+                    )
+                }
+            }
+        },
+        confirmButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(string.close))
+            }
+        }
+    )
+}
